requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
rdkit>=2022.03.1
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.5.0
lightgbm>=3.3.0
torch>=1.10.0
transformers>=4.15.0
shap>=0.40.0
lime>=0.2.0
joblib>=1.1.0
tqdm>=4.62.0
pytest>=6.2.5 